Format and codec changes only
