+++ experimental/dds/tree2/src/feature-libraries/defaultFieldChangeFormat.ts
@@ -4,8 +4,8 @@
  */
 
 import { Static, TSchema, Type } from "@sinclair/typebox";
-import { EncodedJsonableTree } from "../core";
-import { EncodedChangeAtomId } from "./modular-schema";
+import { EncodedJsonableTree, RevisionTagSchema } from "../core";
+import { ChangesetLocalIdSchema } from "./modular-schema";
 
 export const EncodedNodeUpdate = <Schema extends TSchema>(tNodeChange: Schema) =>
 	Type.Union([
@@ -18,7 +18,7 @@ export const EncodedNodeUpdate = <Schema extends TSchema>(tNodeChange: Schema) =
 			 * The node being restored.
 			 */
 			revert: EncodedJsonableTree,
-			changeId: EncodedChangeAtomId,
+			revision: Type.Optional(RevisionTagSchema),
 			changes: Type.Optional(tNodeChange),
 		}),
 	]);
@@ -62,9 +62,9 @@ export type EncodedValueChangeset<Schema extends TSchema> = Static<
 export const EncodedOptionalFieldChange = <Schema extends TSchema>(tNodeChange: Schema) =>
 	Type.Object({
 		/**
-		 * Uniquely identifies the change made to the field.
+		 * Uniquely identifies, in the scope of the changeset, the change made to the field.
 		 */
-		id: EncodedChangeAtomId,
+		id: ChangesetLocalIdSchema,
 		/**
 		 * The new content for the trait. If undefined, the trait will be cleared.
 		 */
@@ -83,7 +83,7 @@ export const EncodedOptionalChangeset = <Schema extends TSchema>(tNodeChange: Sc
 	Type.Object({
 		fieldChange: Type.Optional(EncodedOptionalFieldChange(tNodeChange)),
 		childChange: Type.Optional(tNodeChange),
-		deletedBy: Type.Optional(EncodedChangeAtomId),
+		deletedBy: Type.Optional(RevisionTagSchema),
 	});
 
 export type EncodedOptionalChangeset<Schema extends TSchema> = Static<
