+++ experimental/dds/tree2/src/feature-libraries/defaultFieldChangeTypes.ts
@@ -3,8 +3,8 @@
  * Licensed under the MIT License.
  */
 
-import { ITreeCursorSynchronous, JsonableTree } from "../core";
-import { ChangeAtomId, NodeChangeset } from "./modular-schema";
+import { ITreeCursorSynchronous, JsonableTree, RevisionTag } from "../core";
+import { ChangesetLocalId, NodeChangeset } from "./modular-schema";
 
 export type NodeUpdate =
 	| {
@@ -16,7 +16,7 @@ export type NodeUpdate =
 			 * The node being restored.
 			 */
 			revert: ITreeCursorSynchronous;
-			changeId: ChangeAtomId;
+			revision: RevisionTag | undefined;
 			changes?: NodeChangeset;
 	  };
 
@@ -27,9 +27,17 @@ export interface ValueChangeset {
 
 export interface OptionalFieldChange {
 	/**
-	 * Uniquely identifies the change made to the field.
+	 * Uniquely identifies, in the scope of the changeset, the change made to the field.
+	 * Globally unique across all changesets when paired with the changeset's revision tag.
 	 */
-	readonly id: ChangeAtomId;
+	readonly id: ChangesetLocalId;
+
+	/**
+	 * When populated, indicates the revision that this field change is associated with.
+	 * Is left undefined when the revision is the same as that of the whole changeset
+	 * (which would also be undefined in the case of an anonymous changeset).
+	 */
+	readonly revision?: RevisionTag;
 
 	/**
 	 * The new content for the trait. If undefined, the trait will be cleared.
@@ -54,11 +62,11 @@ export interface OptionalChangeset {
 	childChange?: NodeChangeset;
 
 	/**
-	 * The change that the node `childChange` is referring to was deleted by.
+	 * The revision the node `childChange` is referring to was deleted in.
 	 * If undefined, `childChange` refers to the node currently in this field.
 	 *
 	 * This representation is sufficient for representing changes to the node present before this changeset and
 	 * after this changeset, but not for changes to nodes that existed only transiently in a transaction.
 	 */
-	deletedBy?: ChangeAtomId;
+	deletedBy?: RevisionTag;
 }
