+++ experimental/dds/tree2/src/feature-libraries/modular-schema/fieldChangeHandler.ts
@@ -3,11 +3,11 @@
  * Licensed under the MIT License.
  */
 
-import { FieldKindIdentifier, Delta, FieldKey, Value, TaggedChange, RevisionTag } from "../../core";
-import { Brand, fail, Invariant } from "../../util";
+import { Delta, TaggedChange, RevisionTag } from "../../core";
+import { fail, Invariant } from "../../util";
 import { ICodecFamily, IJsonCodec } from "../../codec";
 import { CrossFieldManager } from "./crossFieldQueries";
-import { ChangesetLocalId } from "./modularSchemaTypes";
+import { ChangesetLocalId, NodeChangeset, RevisionInfo } from "./modularChangeTypes";
 
 /**
  * Functionality provided by a field kind which will be composed with other `FieldChangeHandler`s to
@@ -208,73 +208,6 @@ export type NodeChangeComposer = (changes: TaggedChange<NodeChangeset>[]) => Nod
 export type IdAllocator = (count?: number) => ChangesetLocalId;
 
 /**
- * Changeset for a subtree rooted at a specific node.
- * @alpha
- */
-export interface NodeChangeset extends HasFieldChanges {
-	valueChange?: ValueChange;
-	valueConstraint?: ValueConstraint;
-	nodeExistsConstraint?: NodeExistsConstraint;
-}
-
-/**
- * @alpha
- */
-export interface NodeExistsConstraint {
-	violated: boolean;
-}
-
-/**
- * @alpha
- */
-export interface ValueConstraint {
-	value: Value;
-	violated: boolean;
-}
-
-/**
- * @alpha
- */
-export interface HasFieldChanges {
-	fieldChanges?: FieldChangeMap;
-}
-
-/**
- * @alpha
- */
-export interface ValueChange {
-	/**
-	 * The revision in which this change occurred.
-	 * Undefined when it can be inferred from context.
-	 */
-	revision?: RevisionTag;
-
-	/**
-	 * Can be left unset to represent the value being cleared.
-	 */
-	value?: Value;
-}
-
-/**
- * @alpha
- */
-export interface ModularChangeset extends HasFieldChanges {
-	/**
-	 * The numerically highest `ChangesetLocalId` used in this changeset.
-	 * If undefined then this changeset contains no IDs.
-	 */
-	maxId?: ChangesetLocalId;
-	/**
-	 * The revisions included in this changeset, ordered temporally (oldest to newest).
-	 * Undefined for anonymous changesets.
-	 * Should never be empty.
-	 */
-	readonly revisions?: readonly RevisionInfo[];
-	fieldChanges: FieldChangeMap;
-	constraintViolationCount?: number;
-}
-
-/**
  * A callback that returns the index of the changeset associated with the given RevisionTag among the changesets being
  * composed or rebased. This index is solely meant to communicate relative ordering, and is only valid within the scope of the
  * compose or rebase operation.
@@ -304,42 +237,3 @@ export function getIntention(
 ): RevisionTag | undefined {
 	return rev === undefined ? undefined : revisionMetadata.getInfo(rev).rollbackOf ?? rev;
 }
-
-/**
- * @alpha
- */
-export interface RevisionInfo {
-	readonly revision: RevisionTag;
-	/**
-	 * When populated, indicates that the changeset is a rollback for the purpose of a rebase sandwich.
-	 * The value corresponds to the `revision` of the original changeset being rolled back.
-	 */
-	readonly rollbackOf?: RevisionTag;
-}
-
-/**
- * @alpha
- */
-export type FieldChangeMap = Map<FieldKey, FieldChange>;
-
-/**
- * @alpha
- */
-export interface FieldChange {
-	fieldKind: FieldKindIdentifier;
-
-	/**
-	 * If defined, `change` is part of the specified revision.
-	 * Undefined in the following cases:
-	 * A) A revision is specified on an ancestor of this `FieldChange`, in which case `change` is part of that revision.
-	 * B) `change` is composed of multiple revisions.
-	 * C) `change` is part of an anonymous revision.
-	 */
-	revision?: RevisionTag;
-	change: FieldChangeset;
-}
-
-/**
- * @alpha
- */
-export type FieldChangeset = Brand<unknown, "FieldChangeset">;
